+++ gempak/ush/gdas_ukmet_meta_ver.sh
@@ -49,7 +49,9 @@ for area in ${areas}; do
         sdatenum=${stime:0:6}
         cyclenum=${stime:6}
 
-        ln -sf "${COMINukmet}/ukmet.20${sdatenum}/gempak" "ukmet.20${sdatenum}"
+        if [[ -L "ukmet.20${sdatenum}" ]]; then
+            ln -sf "${COMINukmet}/ukmet.20${sdatenum}/gempak" "ukmet.20${sdatenum}"
+        fi
         gdfile="ukmet.20${sdatenum}/ukmet_20${sdatenum}${cyclenum}${dgdattim}"
 
         # 500 MB HEIGHT METAFILE
